Update imports for renamed use-filter-query-state hook
